Arduino Reddit Scraper + Python Reddit Scraper
Arduino Reddit Scraper:
Added Arduino project which continues off of the Python Reddit Scraper. Allows the user to scrape Reddit articles from a specific sub with the use of buttons and an LCD.

Python Reddit Scraper:
Fixed the issue with the Reddit scraper which would cause the status code of Reddit to be 429 too often. Added a user agent when sending a request.

Also, addeed feature to allow user to delve into a Reddit comment thread of an article.