Nokia Reddit Scraper
Continuation of the 16x2 LCD Reddit Scraper but with a larger LCD screen and a navigable menu.

Uses a Nokia 5110 LCD, a three button setup, and an Arduino Uno.
Followed https://www.youtube.com/watch?v=VtZvf5T98FI as a basis for the menu code.

See readme for more info.