Updates to Nokia Reddit Scraper
Several main updates:
- Added switch-case statements to replace if-else statements when appropriate (had to change var page to be an int)
- Added extra page called "More Options" located in "Options" - contains the Num of Results and EZ Read function
- EZ Read is where words which are split into two on different lines will be syllabified (if possible) so that it is properly divided. If not syllabifiable, then the entire word is shifted to the second line. The Python library "pyphen" is used

Also tinkered with the Adafruit GFX Library. Possible to see on the Contrast page.

Currently the EZ Read has problems on correctly placing the dash on the screen if a previous line had a dash

e.g. Screen will appear like:
The per-
son is eat
-ing

Also EZ Read does not account for multi-syllabic words at the moment
e.g. Screen will appear like
The com-
puter is on.

even though:
The compu-ter
is on.
is a valid printout as well because the first line has sufficient space.