Nokia Reddit Scraper - WiFi
Development of the original Nokia Reddit Scraper code by allowing the user to wirelessly scrape Reddit. It uses an Arduino UNO WiFi to send GET and POST requests to ThingSpeak. The Python script retrieves the scraping parameters from ThingSpeak, scrapes with the parameters, and then posts the results to ThingSpeak. Arduino UNO WiFi then retrieves the data and sends it to the Arduino UNO to display on the LCD. See readme for more details about the entire project.

Other note:
Forgot to upload the updated Nokia Reddit Scraper (original) code. Now allow the user to change the key bindings.